Restructure the repository folders and Docker config
Slack block elements are in modals, functions are in utils, and
each container now has a dedicated folder. This should make it
easier to track files and dependencies.
