+++ app/slack-listener/requirements.txt
@@ -1,0 +1,11 @@
+# Python libraries will be installed when container is built
+# pip install -r requirements.txt
+
+# Slack API
+slackclient>=1.1.0
+slackeventsapi
+slack_bolt
+
+# Python etc.
+python-dotenv
+flask
