+++ docker-compose.yml
@@ -3,20 +3,22 @@ services:
     standard-backend:
         build:
             context: app
-            dockerfile: app.Dockerfile
+            dockerfile: ./app/slack-listener/app.Dockerfile
         env_file:
             - ./app/.env
         ports:
             - "3000:3000"
         volumes:
-            - ./app:/app
+            - ./app/utils:/app/utils:ro
+            - ./app/modals:/app/modals:ro
     group-queue:
         build:
             context: app
-            dockerfile: queue.Dockerfile
+            dockerfile: ./app/grouping-queue/queue.Dockerfile
         env_file:
             - ./app/.env
         ports:
             - "4000:4000"
         volumes:
-            - ./app:/app
+            - ./app/utils:/app/utils:ro
+            - ./app/modals:/app/modals:ro
